% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_meta.R
\name{demo_meta}
\alias{demo_meta}
\title{Demo meta - The demo of meta method}
\usage{
demo_meta(
  A = NULL,
  B,
  s = NULL,
  X = NULL,
  algorithm = c("pearson", "aracne.a", "clr"),
  meta = "union",
  max.steps = 200,
  gene,
  TF = NULL,
  file = NULL,
  cl.cores = 2,
  verbose = TRUE
)
}
\arguments{
\item{A}{Gene expression data of transcriptome factors (i.e. basis function in machine learning).
The class of A are required to be 'matrix' and the dimension of matrix A is m * n.}

\item{B}{Gene expression data of target genes (i.e. observation in machine learning).
The class of B are required to be 'matrix' and the dimension of matrix B is u * n.}

\item{s}{Sparsity level of solution. User can input a sequence of sparsity, i.e. 's <- c(1,2,3,4,5)'.}

\item{X}{Gene expression data of Chromatin immunoprecipitation or zero matrix
(i.e. initial iterative point in machine learning).The class of X are required to be
'matrix' and the dimension of matrix X is u * m.}

\item{algorithm}{The algorithm of inferring grn. We provide 25 algorithms in this function including
'pearson','spearman','kendall','WGCNA','bc3net','c3net','aracne.a','aracne.m',clr','mrnet','mrnetb',
'GeneNet','Genie3.ET','Genie3.RF','pict','ADMM', 'ADMMHalf',ADMMHard', 'ITAHalf', 'ITAHard',
'ISTA', 'SPGL0', 'SPGL1', 'CoSaMP', 'OMP', 'FoBa' and 'Lars'. Defaule: c('pearson','aracne.a','clr')}

\item{meta}{Take the intersection or union of these methods. (union, intersection) Default: union.}

\item{max.steps}{Maximum iteration used in calculation. Default: 200.}

\item{gene}{A vector of characters containing the target genes.}

\item{TF}{The Transcription Factors (TFs) The default value NULL means that all the genes are used as candidate regulators. Default: NULL.}

\item{file}{A connection, or a character string naming the file to write to. The default not save the grn matrix. Default: NULL.}

\item{cl.cores}{The number of cores in computer that you want to utilize. Default: 2.}

\item{verbose}{If set to TRUE, a feedback on the progress of the progress of the calculations is given. Default: TRUE.}
}
\value{
A list with the function of parameters ,algorithms, elapse time, grn which a matrix which is the weighted adjacency matrix of the network inferred by the these algorithms.
}
\description{
This is the main function to call meta algorithms for network inference
}
\details{
The demo function aims to call algorithms to infer gene regulatory network.
For the existing gene regulatory network algorithms, select several algorithms, and take the intersection or union of the regulatory network results obtained by these methods to obtain a new regulatory network result.
}
\examples{
A <- matrix(rnorm(200,0,1),20,10)
B <- matrix(rnorm(1000,0,1),100)
X <- matrix(0,100,20)
s <- c(1:10)
TF <- paste0('G',1:20)
gene <- paste0('G',1:100)
algorithm=c('pearson','aracne.a','clr','OMP')
res <- demo_meta(A=A, B = B,s=s, X=X,algorithm=algorithm,meta='union',max.steps = 200,
                 gene=gene,TF = TF,file = NULL,cl.cores=2,verbose = TRUE)
grn <- res$grn
}
