% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.plot.R
\name{draw.plot}
\alias{draw.plot}
\title{draw.plot}
\usage{
draw.plot(dat, method = "roc", algname)
}
\arguments{
\item{dat}{A data.frame. The first element, dat$predictions, is a vector of numerical predictions. The second element, dat$labels, is a vector of cordatponding class labels.}

\item{method}{Roc curve, pr curve and ss curve(roc,pr,ss). Default: roc.}

\item{algname}{The algorithm of gene network inferring.}
}
\value{
The curve which selected
}
\description{
Draw the curve for the gene regulatory network.
}
\examples{
dat <- data.frame(predictions = runif(200, min=0, max=1),labels = c(rep(1,100),rep(0,100)))
p <- draw.plot(dat = dat,method = 'pr',algname ='ADMM')
}
