% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRN-package.R
\docType{package}
\name{WGRNIA-package}
\alias{WGRNIA}
\alias{WGRNIA-package}
\title{WGRNIA: Wrapping gene regulatory network inference algorithms}
\description{
This package including of several gene network inference algorithms to help users of gene regulatory network construction tools evaluate the performance under experimental data.
}
\keyword{internal}
